+++ javax/src/main/java/me/panpf/javax/util/BoundedQueue.java
@@ -46,7 +46,13 @@ public class BoundedQueue<E> implements Queue<E> {
      * Adjust max size
      */
     public void setMaxSize(int maxSize) {
-        Premisex.require(maxSize >= 1, "maxSize not less than 1");
+        Premisex.require(maxSize >= 1, new LazyValue<String>() {
+            @NotNull
+            @Override
+            public String get() {
+                return "maxSize not less than 1";
+            }
+        });
         this.maxSize = maxSize;
         adjustSize(0);
     }